backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.9
celery==5.4.0
redis==5.0.4
boto3==1.34.107
jinja2==3.1.4
openpyxl==3.1.2
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
